+++ inside-java/README.adoc
