+++ inside-java/build.gradle.kts
