+++ inside-java/src/main/kotlin/by/jprof/telegram/opinions/insidejava/config/env.kt
