+++ inside-java/src/main/kotlin/by/jprof/telegram/opinions/insidejava/dao/InsideJavaDAO.kt
