+++ inside-java/src/main/kotlin/by/jprof/telegram/opinions/insidejava/entity/InsideJava.kt
