+++ settings.gradle.kts
@@ -1,6 +1,6 @@
 rootProject.name = "opinions-bot"
 
 include(":webhook")
-include(":inside-java-podcast-poster")
+include(":inside-java")
 include(":components")
 include(":news-queue")
